refactor(dev): extract ADR-0018 fault domain into devcake_dev package
Phase 1 of the Dev entrypoint split: pure harness fault classification lives
in images/common/devcake_dev/domain/fault.py (Zone B package, not app hexagon).
/dev_entrypoint.py stays the ENTRYPOINT façade and re-exports the public API
so existing importlib tests keep working. Dockerfile copies the package to
/devcake_dev; docs/07 notes the layout.
