+++ images/common/devcake_dev/__init__.py
@@ -1,0 +1,4 @@
+"""DevCake Dev process package (Zone B) — hexagonal layout for the shared entrypoint.
+
+Control plane lives in app/devcake. This package is copied only into Dev images.
+"""
