+++ images/common/devcake_dev/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Pure Dev-side domain (no Redis / subprocess)."""
